refactor: extract DomainError and assertAdminUser into shared module
